Move GenerateCommandErrorOutputFactory
